scripts/parser_server/requirements.txt
Flask
flask-cors
pdfminer.six
python-docx
requests
